backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
supabase==2.3.1
pytest==8.3.5
pytest-mock==3.14.0
python-multipart==0.0.9
xmltodict==0.13.0
